@@ -28,36 +28,36 @@
   "license": "MIT",
   "schematics": "./src/collection.json",
   "dependencies": {
-    "@angular-devkit/core": "^20.3.15",
-    "axios": "^1.13.2",
+    "axios": "^1.13.5",
     "lodash": "^4.17.23",
     "pluralize": "^8.0.0",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@angular-devkit/schematics": "^20.3.15",
-    "@angular-devkit/schematics-cli": "^21.1.1",
-    "@angular/compiler": "^21.1.1",
-    "@angular/core": "^21.1.1",
-    "@eslint/js": "^9.39.2",
+    "@angular-devkit/core": "^21.1.4",
+    "@angular-devkit/schematics": "^21.1.4",
+    "@angular-devkit/schematics-cli": "^21.1.4",
+    "@angular/compiler": "^21.1.4",
+    "@angular/core": "^21.1.4",
+    "@eslint/js": "^10.0.1",
     "@jest/globals": "^30.2.0",
     "@types/lodash": "^4.17.23",
-    "@types/node": "^22.19.7",
+    "@types/node": "^24.10.13",
     "@types/pluralize": "^0.0.33",
-    "@typescript-eslint/eslint-plugin": "^8.53.1",
-    "@typescript-eslint/parser": "^8.8.1",
-    "@typescript-eslint/typescript-estree": "^8.8.1",
-    "eslint": "^9.39.2",
+    "@typescript-eslint/eslint-plugin": "^8.55.0",
+    "@typescript-eslint/parser": "^8.55.0",
+    "@typescript-eslint/typescript-estree": "^8.55.0",
+    "eslint": "^10.0.1",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-angular": "^5.0.0",
-    "eslint-plugin-jest": "^29.12.1",
+    "eslint-plugin-jest": "^29.14.0",
     "eslint-plugin-jest-dom": "^5.5.0",
     "eslint-plugin-jest-formatting": "^3.1.0",
     "eslint-plugin-ngrx": "^2.1.4",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^5.5.5",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "eslint-plugin-sonarjs": "^3.0.5",
+    "eslint-plugin-sonarjs": "^4.0.0",
     "eslint-plugin-standard": "^4.1.0",
     "eslint-plugin-typescript": "^0.14.0",
     "jest": "^30.2.0",
@@ -67,6 +67,6 @@
     "snyk": "^1.1302.1",
     "ts-jest": "^29.4.6",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^8.53.1"
+    "typescript-eslint": "^8.55.0"
   }
 }
