@@ -10,6 +10,7 @@ import { plural, singular } from 'pluralize';
 import { classify, dasherize } from '@angular-devkit/core/src/utils/strings';
 
 const collectionPath = path.join(__dirname, `../collection.json`);
+const thirtySeconds = 30000;
 
 describe(`imng-module AdventureWorks Orders`, () => {
   let tree: UnitTestTree;
@@ -25,7 +26,7 @@ describe(`imng-module AdventureWorks Orders`, () => {
       modelFolderLocation: '../../../models/webapi',
     };
     tree = await runner.runSchematic(`imng-ngrx-module`, options, Tree.empty());
-  }, 30000);
+  }, thirtySeconds);
 
   test(`tree files should match`, () => {
     expect(tree.files.sort()).toMatchSnapshot();
@@ -33,7 +34,7 @@ describe(`imng-module AdventureWorks Orders`, () => {
 
   test(`module should work`, () => {
     const file = tree.get(
-      `/test/orders-ngrx-module/orders.module.ts`
+      `/test/orders-ngrx-module/orders-ngrx.module.ts`
     );
     const content = file?.content.toString();
     expect(content).toMatchSnapshot();
