@@ -9,6 +9,7 @@ async function generateTsModels() {
       'https://awod-ikemtz.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/adventure-works/src/app/models/odata',
     genAngularFormGroups: true,
+    genAngularFormGroupsWithDefaultValues: true,
     typeFilterCallBack: generator.nrsrxTypeFilterCallBack,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
@@ -28,6 +29,7 @@ async function generateTsModels() {
       'https://awwa-ikemtz.azurewebsites.net/swagger/v1/swagger.json',
     outputPath: './apps/adventure-works/src/app/models/webapi',
     genAngularFormGroups: true,
+    genAngularFormGroupsWithDefaultValues: true,
     valuePropertyTypeFilterCallBack: (val, i, arr) =>
       !val.name.startsWith('created') && !val.name.startsWith('updated'),
     pathUrlFormattingCallBack: (val) =>
@@ -41,4 +43,4 @@ async function generateTsModels() {
   });
 }
 
-generateTsModels();
+await generateTsModels();
