@@ -19,36 +19,29 @@ import { IOrderLineItemForm } from './order-line-item.form';
 
 export function OrderFormGroupFac(): FormGroup<IOrderForm> {
   return new FormGroup<IOrderForm>({
-    id: new FormControl<string | null | undefined>(undefined),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    orderId: new FormControl<number>(undefined, {
+    id: new FormControl<string | null | undefined>(null),
+    orderId: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    revisionNum: new FormControl<number>(undefined, {
+    revisionNum: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    date: new FormControl<Date>(undefined, {
+    date: new FormControl<Date>(new Date(), {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    dueDate: new FormControl<Date>(undefined, {
+    dueDate: new FormControl<Date>(new Date(), {
       validators: Validators.required,
       nonNullable: true,
     }),
-
-    shipDate: new FormControl<Date | null | undefined>(undefined),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    isOnlineOrder: new FormControl<boolean>(undefined, {
+    shipDate: new FormControl<Date | null | undefined>(null),
+    isOnlineOrder: new FormControl<boolean>(false, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    num: new FormControl<string>(undefined, {
+    num: new FormControl<string>('', {
       validators: Validators.compose([
         Validators.required,
         Validators.minLength(1),
@@ -56,55 +49,42 @@ export function OrderFormGroupFac(): FormGroup<IOrderForm> {
       ]),
       nonNullable: true,
     }),
-
-    purchaseOrderNum: new FormControl<string | null | undefined>(undefined, {
+    purchaseOrderNum: new FormControl<string | null | undefined>(null, {
       validators: Validators.maxLength(25),
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    customerId: new FormControl<string>(undefined, {
+    customerId: new FormControl<string>('', {
       validators: Validators.required,
       nonNullable: true,
     }),
-
-    shipToAddressId: new FormControl<string | null | undefined>(undefined),
-
-    billToAddressId: new FormControl<string | null | undefined>(undefined),
-
-    creditCardApprovalCode: new FormControl<string | null | undefined>(
-      undefined,
-      { validators: Validators.maxLength(15) },
-    ),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    subTotal: new FormControl<number>(undefined, {
+    shipToAddressId: new FormControl<string | null | undefined>(null),
+    billToAddressId: new FormControl<string | null | undefined>(null),
+    creditCardApprovalCode: new FormControl<string | null | undefined>(null, {
+      validators: Validators.maxLength(15),
+    }),
+    subTotal: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    taxAmt: new FormControl<number>(undefined, {
+    taxAmt: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    freight: new FormControl<number>(undefined, {
+    freight: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    totalDue: new FormControl<number>(undefined, {
+    totalDue: new FormControl<number>(0, {
       validators: Validators.required,
       nonNullable: true,
     }),
-
-    comment: new FormControl<string | null | undefined>(undefined, {
+    comment: new FormControl<string | null | undefined>(null, {
       validators: Validators.maxLength(500),
     }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    statusType: new FormControl<OrderStatusTypes | string>(undefined, {
-      validators: Validators.required,
-      nonNullable: true,
-    }),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    shippingType: new FormControl<ShippingTypes | string>(undefined, {
+    statusType: new FormControl<OrderStatusTypes | string>(
+      OrderStatusTypes.Processing,
+      { validators: Validators.required, nonNullable: true },
+    ),
+    shippingType: new FormControl<ShippingTypes | string>(ShippingTypes.Other, {
       validators: Validators.required,
       nonNullable: true,
     }),
