@@ -13,9 +13,8 @@ import { IProductForm } from './product.form';
 
 export function ProductModelFormGroupFac(): FormGroup<IProductModelForm> {
   return new FormGroup<IProductModelForm>({
-    id: new FormControl<string | null | undefined>(undefined),
-    // @ts-expect-error this is intentional, we want the initial value to be undefined so that the required validator will work correctly.
-    name: new FormControl<string>(undefined, {
+    id: new FormControl<string | null | undefined>(null),
+    name: new FormControl<string>('', {
       validators: Validators.compose([
         Validators.required,
         Validators.minLength(1),
@@ -23,8 +22,7 @@ export function ProductModelFormGroupFac(): FormGroup<IProductModelForm> {
       ]),
       nonNullable: true,
     }),
-
-    description: new FormControl<string | null | undefined>(undefined, {
+    description: new FormControl<string | null | undefined>(null, {
       validators: Validators.maxLength(500),
     }),
     products: new FormArray<FormGroup<IProductForm>>([]),
