@@ -1,5 +1,11 @@
-export * from './+state';
-export { CustomersModule } from './customers.module';
+export {
+  customerActionTypes,
+  customersFeature,
+  CustomerListEffects,
+  CustomerLookupEffects,
+  CustomerCrudEffects,
+} from './+state';
+export { CustomersNgrxModule } from './customers-ngrx.module';
 export { CustomerApiService } from './customer.api.service';
 export { CustomerCrudFacade } from './customer.crud.facade';
 export { CustomerListFacade } from './customer.list.facade';
