@@ -1,7 +1,11 @@
-export * from './lib/kendo-odata-component-base';
-export * from './lib/kendo-odata-grid-facade';
-export * from './lib/kendo-odata-grid-state';
-export * from './lib/kendo-odata-grid-rxjs-operators';
-export * from './lib/kendo-odata-initial-state-creator';
-export * from './lib/kendo-odata-grid.module';
-export * from './lib/kendo-odata-grid.directive';
+export { KendoODataBasedComponent } from './lib/kendo-odata-component-base';
+export { IKendoODataGridFacade } from './lib/kendo-odata-grid-facade';
+export { KendoODataGridState } from './lib/kendo-odata-grid-state';
+export {
+  getODataPagerSettings,
+  mapPagerSettings,
+} from './lib/kendo-odata-grid-rxjs-operators';
+export { createKendoODataGridInitialState } from './lib/kendo-odata-initial-state-creator';
+export { ImngKendoGridODataModule } from './lib/kendo-odata-grid.module';
+export { IMNG_KENDO_GRID_ODATA } from './lib/kendo-odata-grid.directive';
+export { IMNG_KENDO_GRID_ODATA_HEADER } from './lib/grid-odata-header-component/grid-odata-header.component';
