@@ -3,6 +3,7 @@ import { CommonModule } from '@angular/common';
 import { IMNG_KENDO_GRID_ODATA } from './kendo-odata-grid.directive';
 import { ExcelModule, PDFModule } from '@progress/kendo-angular-grid';
 import { ImngKendoGridModule } from 'imng-kendo-grid';
+import { IMNG_KENDO_GRID_ODATA_HEADER } from './grid-odata-header-component/grid-odata-header.component';
 
 @NgModule({
   imports: [
@@ -11,7 +12,8 @@ import { ImngKendoGridModule } from 'imng-kendo-grid';
     ExcelModule,
     ImngKendoGridModule,
     IMNG_KENDO_GRID_ODATA,
+    IMNG_KENDO_GRID_ODATA_HEADER,
   ],
-  exports: [IMNG_KENDO_GRID_ODATA],
+  exports: [IMNG_KENDO_GRID_ODATA, IMNG_KENDO_GRID_ODATA_HEADER],
 })
-export class ImngKendoGridODataModule {}
+export class ImngKendoGridODataModule { }
